nginx_exporter_metrics: major refactor before 1.8.2
  * rename plugin to in_nginx_exporter_metrics
  * change some values to use a gauge instead of
    a counter.
  * refactor all the separate config (ctx) code
    into the main file.
  * Use uin64_t values for the metrics.
  * Add new gauge value that is set to 1.0 when the
    nginx stub_status module answers correctly and
    is set to 0.0 when it does not.
