+++ CMakeLists.txt
@@ -128,7 +128,7 @@ option(FLB_IN_HEAD            "Enable Head input plugin"            Yes)
 option(FLB_IN_PROC            "Enable Process input plugin"         Yes)
 option(FLB_IN_SYSTEMD         "Enable Systemd input plugin"         Yes)
 option(FLB_IN_DUMMY           "Enable Dummy input plugin"           Yes)
-option(FLB_IN_NGINX_STUB_STATUS    "Enable Nginx status input plugin"    Yes)
+option(FLB_IN_NGINX_EXPORTER_METRICS "Enable Nginx Metrics input plugin" Yes)
 option(FLB_IN_NETIF           "Enable NetworkIF input plugin"       Yes)
 option(FLB_IN_WINLOG          "Enable Windows Log input plugin"      No)
 option(FLB_IN_WINSTAT         "Enable Windows Stat input plugin"     No)
