+++ plugins/in_nginx_exporter_metrics/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(src
+  nginx.c)
+
+FLB_PLUGIN(in_nginx_exporter_metrics "${src}" "")
