--- plugins/in_nginx_stub_status/CMakeLists.txt
@@ -1,5 +1,0 @@
-set(src
-  nginx_stub_status.c
-  nginx_stub_status_config.c)
-
-FLB_PLUGIN(in_nginx_stub_status "${src}" "")
